+++ example/order_scope_level/feature1/test_a.py
@@ -1,0 +1,11 @@
+import pytest
+
+
+@pytest.mark.order(2)
+def test_two():
+    pass
+
+
+@pytest.mark.order(1)
+def test_one():
+    pass